% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.v16.R
\name{find_starts}
\alias{find_starts}
\title{Detect start of waves}
\usage{
find_starts(rt_values, peaks_x, search_range)
}
\arguments{
\item{rt_values}{A vector containing rt values for each time point.}

\item{peaks_x}{Time points of peaks based on daily cases (one for each wave).}

\item{search_range}{The range of data points to go through for filtering invalid beginnings (a vector).}
}
\value{
A list containing the detected start times of waves, as scalars.
}
\description{
Find the approximate beginning times of waves given times of peaks. This is based on local maxima of \eqn{R_t}.
}
\details{
Note: This is provided for convenience only, and is not meant to replace an analyst's determination of wave bounds.
}
