% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executions.v16.R
\name{load_config}
\alias{load_config}
\title{Load configuration file}
\usage{
load_config()
}
\value{
A ’config’ object, which is a list that stores input parameters and
   settings from config.yml. Variable names are imported exactly. See the
   configuration file for details.
}
\description{
Load the configuration file as an object in the global
   environment. This can be passed to the main functions instead of the
   individual arguments within, for user convenience.
}
