% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.v16.R
\name{modif.helper}
\alias{modif.helper}
\title{Helper function which ensures that parameters are within specified bounds. Called by the estimate.mle.
Note: this is an internal function which should not be modified.}
\usage{
modif.helper(params_modif, params_limits)
}
\arguments{
\item{params_modif}{Raw parameter vector, in the same order as argument ini_param in estimate.mle.}

\item{params_limits}{Boundaries/limits of the ini_params.}
}
\value{
The modified parameters (bounded, if needed).
}
\description{
Helper function which ensures that parameters are within specified bounds. Called by the estimate.mle.
Note: this is an internal function which should not be modified.
}
\keyword{internal}
