% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executions.v16.R
\name{navigate_to_config}
\alias{navigate_to_config}
\title{Navigate to the config file}
\usage{
navigate_to_config()
}
\value{
The path to the configuration file.
}
\description{
Prints the path to the config file and opens the config file.
}
