% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.v16.R
\name{ranges_to_waves}
\alias{ranges_to_waves}
\title{Utility function for range manipulation}
\usage{
ranges_to_waves(waves_list)
}
\arguments{
\item{waves_list}{A list containing ranges (start, end) of each wave.}
}
\value{
A two-dimensional list with individual waves as sub-lists.
}
\description{
Converts a list of waves to a two-dimensional list.
}
