% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executions.v16.R
\name{re_predict}
\alias{re_predict}
\title{Demo of main functions}
\arguments{
\item{path_to_data}{Absolute path to the dataset in csv format.}
}
\value{
No return value.
}
\description{
Fits the model to an example case data and predicts the epidemic curves and plots the outputs.
}
\details{
Please modify the config file before invoking this method.
 The config file contains all the settings and initial  parameter values
 necessary for the algorithm to run. Path to the dataset (in csv format)
 is also set in the config file. This file should be updated to the desired
 specifications before running this demo. To open it, execute
 REffectivePred::navigate_to_config().
}
