% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.v16.R
\name{serial.helper}
\alias{serial.helper}
\title{Serial interval}
\usage{
serial.helper(window_size, serial_mean = log(4), serial_var = log(1.380715))
}
\arguments{
\item{window_size}{The maximum value for the serial interval.}

\item{serial_mean}{Mean of the serial interval on the log scale. See Details.}

\item{serial_var}{Variance of the serial interval on the log scale. See Details.}
}
\value{
A vector that stores the serial interval in reverse order. This is
 meant to be multiplied to infections in chronological order.
}
\description{
Helper function for computing weights \code{w} based on the serial interval.
}
\details{
Computed based on a log normal density function, as in Nishiura et al. (2020). Parameters
 \code{serial_mean} and \code{serial_var} are arguments \code{meanlog} and \code{sdlog} of
 function \code{dlnorm}. Default values are taken from the same reference.
}
