% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.v16.R
\name{waves_1d_list}
\alias{waves_1d_list}
\title{Utility function for range manipulation}
\usage{
waves_1d_list(num_waves, waves_list)
}
\arguments{
\item{num_waves}{Total number of waves.}

\item{waves_list}{A list containing individual waves as sub-lists.}
}
\value{
A vector with the first \code{num_waves} waves combined.
}
\description{
Combines multiple waves into one vector.
}
